refactor: v0.4.0-SNAPSHOT

+++ src/main/kotlin/space/iseki/cmdpipe/ExecutionInfo.kt
@@ -3,7 +3,7 @@ package space.iseki.cmdpipe
 import java.time.Instant
 
 data class ExecutionInfo internal constructor(
-    val pid: Int,
+    val pid: Long,
     val startAt: Instant,
     val endAt: Instant? = null,
     val exitCode: Int? = null,
